pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "phishnet"
version = "0.3.0"
description = "Real-time phishing URL classifier with drift detection"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "Haya Alshammari", email = "hayambg@gmail.com"}]
dependencies = [
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.0.0",
    "onnxruntime>=1.16.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0", "pytest-cov>=4.1.0",
    "skl2onnx>=1.15.0", "httpx>=0.24.0",
    "evidently>=0.4.0", "matplotlib>=3.7.0",
    "mlflow>=2.8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
